dev_requirements.txt
asgiref==3.3.1
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
config==0.5.0.post0
decorator==4.4.2
Django==3.1.6
django-autoslug==1.9.8
django-bootstrap-form==3.4
django-bootstrap-v5==1.0.4
django-extensions==3.1.1
django-leaflet==0.27.1
django-rosetta==0.9.6
djangorestframework==3.12.4
gunicorn==20.1.0
idna==2.10
importlib-metadata==3.10.1
iniconfig==1.1.1
ipython==7.21.0
ipython-genutils==0.2.0
jedi==0.18.0
jsonfield==3.1.0
packaging==21.0
parso==0.8.1
pickleshare==0.7.5
Pillow==8.1.0
pluggy==0.13.1
polib==1.1.1
prompt-toolkit==3.0.16
psycopg2-binary==2.8.6
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
python-decouple==3.4
python-slugify==4.0.1
pytz==2021.1
requests==2.25.1
selenium==3.141.0
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.6
uWSGI==2.0.19.1
wcwidth==0.2.5
zipp==3.4.1
